+++ app/src/main/java/ru/hse/routemood/auth/AuthController.java
@@ -1,10 +1,8 @@
 package ru.hse.routemood.auth;
 
-import java.util.List;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,8 +11,6 @@ import ru.hse.routemood.auth.domain.dto.AuthRequest;
 import ru.hse.routemood.auth.domain.dto.AuthResponse;
 import ru.hse.routemood.auth.domain.dto.RefreshRequest;
 import ru.hse.routemood.auth.domain.dto.RegisterRequest;
-import ru.hse.routemood.auth.domain.models.User;
-import ru.hse.routemood.auth.repository.UserServiceRepository;
 import ru.hse.routemood.auth.services.AuthService;
 
 @RestController
@@ -22,8 +18,6 @@ import ru.hse.routemood.auth.services.AuthService;
 @AllArgsConstructor
 public class AuthController {
 
-
-    private final UserServiceRepository userServiceRepository;
     private final AuthService authService;
 
     @PostMapping(path = "/register")
@@ -56,11 +50,5 @@ public class AuthController {
         }
 
         return ResponseEntity.ok(result);
-    }
-
-    @GetMapping(path = "/users")
-    public ResponseEntity<List<User>> listUsers() {
-        List<User> users = userServiceRepository.findAll();
-        return ResponseEntity.ok(users);
     }
 }
