+++ app/src/main/java/ru/hse/routemood/auth/domain/dto/RefreshRequest.java
@@ -10,6 +10,7 @@ import lombok.Setter;
 @AllArgsConstructor
 @Builder
 public class RefreshRequest {
+
     private String accessToken;
     private String refreshToken;
     private String username;
