+++ app/src/main/java/ru/hse/routemood/auth/services/JwtService.java
@@ -12,7 +12,7 @@ import javax.crypto.SecretKey;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
-import ru.hse.routemood.auth.domain.models.User;
+import ru.hse.routemood.user.domain.models.User;
 
 @Service
 public class JwtService {
