+++ app/src/main/java/ru/hse/routemood/user/repository/UserServiceRepository.java
@@ -1,9 +1,9 @@
-package ru.hse.routemood.auth.repository;
+package ru.hse.routemood.user.repository;
 
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.hse.routemood.auth.domain.models.User;
+import ru.hse.routemood.user.domain.models.User;
 
 public interface UserServiceRepository extends JpaRepository<User, UUID> {
